Create MainFragment and recreate same funcionality in MainFragment from MainActivity
