+++ app/src/main/res/layout/activity_main.xml
@@ -1,75 +1,19 @@
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
-    android:orientation="vertical"
-    android:gravity="center"
-    android:layout_height="match_parent">
+    android:layout_height="match_parent"
+    android:orientation="vertical">
 
-    <!--<include layout="@layout/base_toolbar"/>-->
-    <android.support.design.widget.CoordinatorLayout
-            android:id="@+id/myCoordinatorLayout"
-            android:layout_width="match_parent"
-            android:gravity="center"
-            android:layout_height="match_parent"
-            >
+    <android.support.design.widget.AppBarLayout
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        app:elevation="0dp"
+        android:elevation="0dp"
+        >
 
-        <LinearLayout
-            android:id="@+id/toDoEmptyView"
-            android:orientation="vertical"
-            android:gravity="center"
-            android:layout_width="match_parent"
-            android:layout_height="match_parent">
-
-            <ImageView
-                android:src="@drawable/empty_view_bg"
-                android:layout_width="100dp"
-                android:layout_height="100dp" />
-            <TextView
-                android:text="@string/no_to_dos"
-                android:textColor="@color/secondary_text"
-                android:textSize="16sp"
-                android:paddingTop="4dp"
-                android:paddingBottom="8dp"
-                android:gravity="center"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content" />
-
-        </LinearLayout>
-
-        <android.support.design.widget.AppBarLayout
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            app:elevation="0dp"
-            android:elevation="0dp"
-            >
-
-            <include layout="@layout/base_toolbar"/>
-        </android.support.design.widget.AppBarLayout>
-
-            <!--<include layout="@layout/base_toolbar"/>-->
-
-
-        <!--</android.support.design.widget.AppBarLayout>-->
-
-
-            <com.example.avjindersinghsekhon.minimaltodo.Utility.RecyclerViewEmptySupport
-                app:layout_behavior="@string/appbar_scrolling_view_behavior"
-                android:id="@+id/toDoRecyclerView"
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"/>
-
-
-        <android.support.design.widget.FloatingActionButton
-            android:src="@drawable/ic_add_white_24dp"
-            android:id="@+id/addToDoItemFAB"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_margin="16dp"
-            app:layout_anchor="@id/myCoordinatorLayout"
-            app:layout_anchorGravity="bottom|right|end"
-            app:layout_behavior="com.example.avjindersinghsekhon.minimaltodo.ScrollingFABBehaviour"
-            />
-
-
-    </android.support.design.widget.CoordinatorLayout>
+        <include layout="@layout/base_toolbar"/>
+        <include layout="@layout/layout_fragment_container"/>
+    </android.support.design.widget.AppBarLayout>
 </LinearLayout>
